Comprehensive audit: fix 14 bugs, add missing DB indexes, .gitignore, and improvement backlog
Bug fixes across all layers:
- processing/rtl_wideband.py: parse peak_threshold as float (int crashed on negative dBm)
- processing/enrich_wideband.py: guard division-by-zero in avg_power calc; replace bare
  except with specific exception types to stop swallowing KeyboardInterrupt
- processing/enrich.py: add 5 missing indexes to schema DDL (wifi_obs/rf_obs timestamp,
  geo partial indexes on all three obs tables); add WAL checkpoint(TRUNCATE) after each run
- webapp/main.py: add _sudoers_snippet() helper and wire into both sudo error responses so
  the collectors modal shows actionable config instead of undefined; fix nmea_to_dd crash on
  malformed NMEA missing decimal point; use consistent pgrep -f flag; replace date regex
  with datetime.strptime() to reject calendar-invalid dates (2024-13-32)
- webapp/manage.sh: move PID file and log out of world-writable /tmp to project-local paths
- webapp/static/live.js: guard counter increment against unknown observation type
- wardrive.sh: add SDR frequency range validation (24-1766 MHz) before starting rtl_433;
  scale gps_query_fix per-attempt timeout with GPS_WAIT_FIX instead of hardcoded 8s

Also adds:
- .gitignore (excludes *.db, __pycache__, capture/raw/, capture/logs/, logs, pid files)
- README.md: fix stale /tmp log path and /bin/kill sudoers references; add Improvement
  Backlog section with prioritized software, hardware, and runtime recommendations

+++ processing/rtl_wideband.py
@@ -198,7 +198,7 @@ def main():
     end_mhz        = int(sys.argv[3]) if len(sys.argv) > 3 else 1700  # R820T upper limit ~1766 MHz
     step_mhz       = int(sys.argv[4]) if len(sys.argv) > 4 else 2
     scan_time      = int(sys.argv[5]) if len(sys.argv) > 5 else 1     # seconds integration per step
-    peak_threshold = int(sys.argv[6]) if len(sys.argv) > 6 else -40
+    peak_threshold = float(sys.argv[6]) if len(sys.argv) > 6 else -40.0
 
     scanner = WidebandScanner(output_dir, start_mhz, end_mhz, step_mhz, scan_time,
                               peak_threshold)
